backend/python/requirements.txt
# Core Dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# Database
supabase==2.3.4
postgrest==0.13.2

# Google Gemini AI
google-generativeai==0.3.2

# Security & Authentication
bcrypt==4.1.2
pyjwt==2.8.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0